requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
databricks-connect==15.4.7
databricks-sdk==0.46.0
fonttools==4.56.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==24.2
protobuf==5.29.4
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
rsa==4.9
seaborn==0.13.2
setuptools==77.0.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
uv==0.6.8
